Services/LLM/requirements.txt
# --- Base API ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1

# --- LangChain (importante: alineadas entre sí) ---
langchain==0.3.7
langchain-openai==0.2.2
langchain-community==0.3.2
openai==1.51.0


# --- MongoDB ---
pymongo>=4.7.0
motor>=3.5.0

# --- Networking ---
httpx==0.27.2
requests>=2.32.3

# --- Datos ---
pydantic>=2.7.4
pydantic-settings>=2.3.0

# --- Logging ---
structlog>=24.1.0

# --- Testing ---
pytest>=8.2.0
pytest-asyncio>=0.23.5
